requirements.txt
requests
pillow
pytesseract
sentence-transformers
numpy
torch
fastapi
